+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 --------
+## [unreleased] - 
+### Added
+* Added project/venue support [5](https://github.com/unity-sds/unity-py/issues/58)
+### Fixed 
+### Changed
+### Removed
+### Security
+### Deprecated
+
 
 ## [0.2.1] - 2023-11-29
 ### Added